backend/requirements-cloud.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
cryptography>=42.0.0
python-multipart>=0.0.9
openai>=1.106.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
google-genai>=1.0.0
azure-ai-inference>=1.0.0b9
azure-core>=1.35.0
httpx>=0.28.0
aiohttp>=3.12.0
python-dotenv>=1.1.0
PyYAML>=6.0.2
requests>=2.32.0
pyjwt>=2.8.0
argon2-cffi>=23.1.0
